% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromploid.R
\docType{package}
\name{chromploid}
\alias{chromploid}
\alias{chromploid-package}
\title{chromploid A package that calculates likelihood for chromosome number or ploidy change models in a phylogeny.}
\description{
The chromploid package builds Q matrices that allow for calculation of likelihoods and profile likelihoods for parameters of chromosomal and ploidy change
}
\section{chromploid functions}{

negloglikelihood
Q_bichrom  
Q_bibichrom
Q_unibichrom
Q_reducedbichrom
bichrom_dataset
Q_ploidevol
Q_uniploidevol
Q_biploidevol
ploidevol_dataset
stationary_ploidevol
negloglikelihood
}

