% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromploid_nllike.R
\name{chromploid_nllike}
\alias{chromploid_nllike}
\title{Calculates the negative log-likelihood of a chromosome or ploidy model}
\usage{
chromploid_nllike(log.par, phy, tip.values, pi.0, Q.FUN, Q.ARGS)
}
\arguments{
\item{log.par}{A vector of parameter values in ln scale. Length of the vector depends on model defined by Q.FUN}

\item{phy}{Binary rooted phylogenetic tree}

\item{tip.values}{A data frame with at least two columns first column has tip labels in the same order than phylogenetic tree, second column ploidy or chromosome numbers depending on Q.FUN}

\item{pi.0}{Probability distribution to be used at the root of phylogeny}

\item{Q.FUN}{A function that calculates the Q-matrix of the parameters of interest}

\item{Q.ARGS}{A list of arguments to be passed to Q.FUN}
}
\value{
Result is the negative log-likelihood of parameters log.par given the model in Q.FUN
}
\description{
Calculates the negative log-likelihood of a chromosome or ploidy model
}
\author{
Rosana Zenil-Ferguson
}
