% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromploid_optim.R
\name{chromploid_optim}
\alias{chromploid_optim}
\title{Calculates the maximum likelihood estimate and the value of the likelihood at the MLE of a chromosome or ploidy model (wrapper function of nloptr for chromploid models)
For now it only works for bichrom and bichrom reduced model}
\usage{
chromploid_optim(phy.tree, tip.values, model = "bichrom",
  model.args = list(size = 50), starting.val, root.dist,
  optim.options = list(algorithm = "NLOPT_LN_SBPLX", ftol_rel = 1e-08,
  print_level = 1, maxtime = 1.7e+08, maxeval = 1000))
}
\arguments{
\item{phy.tree}{phylogenetic tree}

\item{tip.values}{A vector with chromosome numbers and/or binary trait states. See Zenil-Ferguson et al 2017 supplement for an explanation of BichroM model.}

\item{model}{options are "bichrom" or "reducedbichrom"}

\item{model.args}{List of arguments passed in each specified model other than parameters}

\item{starting.val}{vector of parameters in log scale that help with the optimization of the likelihood}

\item{root.dist}{vector of probabilities for the states at the root}

\item{optim.options}{See opts from nloptr package, optimization options for the likelihood function}
}
\description{
Calculates the maximum likelihood estimate and the value of the likelihood at the MLE of a chromosome or ploidy model (wrapper function of nloptr for chromploid models)
For now it only works for bichrom and bichrom reduced model
}
\author{
Rosana Zenil-Ferguson
}
