% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidevol_dataset.R
\name{ploidevol_dataset}
\alias{ploidevol_dataset}
\title{Cleans and organizes dataset for PloidEvol analyses}
\usage{
ploidevol_dataset(ploidies, size = 12)
}
\arguments{
\item{ploidies}{a vector with ploidy values, and taxon row.names}

\item{size}{maximum ploidy value allow for analyses, by default is 12. See details for more information.}
}
\value{
ploidy.dataset a data frame whose first column is taxon name and second column the ploidy values-1. Saves the data frame as .txt file for future references. See details.
}
\description{
Cleans and organizes dataset for PloidEvol analyses
}
\details{
This function cleans and organizes a vector of ploidies assuming ploidy values range from 2x to 12x. It deletes values larger than 12x and redefines ploidy values as value-1 since Q_ploidevol indices range from 1 to 11. If larger ploidy values are needed user needs to specify own Q.FUN to use in negloglikelihood.
}
\seealso{
\code{\link{negloglikelihood}}
}

