% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_bichromlinear.R
\name{Q_bichromlinear}
\alias{Q_bichromlinear}
\title{Calculates Q-matrix for BiChroM model}
\usage{
Q_bichromlinear(log.theta, size)
}
\arguments{
\item{log.theta}{vector of size 12 indicating parameters in ln scale for BiChoM the order of the parameters is (lambda0, lambda1, mu0, mu1, rho0, rho1, kappa0, q01, kappa1, q10, e0, e1)}

\item{size}{Maximum number of chromosomes in the sample (recommended no more than 50, states larger than that should be coded as 51)}
}
\value{
Q a sparse matrix of size 2*(size+1)
}
\description{
Calculates Q-matrix for BiChroM model
}
\details{
Q_bichromlinear determines the Q-matrix for a model of chromosome number change associated with a binary state with traits coded as 0 or 1. It is linear because binary trait change rate is a linear function of chromosome number.

Note: This is a beta version, the linear function for binary trait change rate from state 0 to 1 is  kappa0+ (chromosome number)*q01
}
