% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_biploidevol.R
\name{Q_biploidevol}
\alias{Q_biploidevol}
\title{Calculates Q-matrix for bivariate profile likelihoods for Ploidevol model}
\usage{
Q_biploidevol(log.pars, log.theta0, param.name)
}
\arguments{
\item{log.pars}{vector of size 4 indicating values in ln of nuisance parameters (excluding the ones of interest for the profile likelihood)}

\item{log.theta0}{vector of size 2 indicating values parameters of interest in BiChroM. Options are (alpha, beta, delta, epsilon, rho, omega)}

\item{param.name}{vector of size 2 indicating the names of parameters of interest. E.g. c("alpha",beta")}
}
\value{
Q a sparse matrix of size 11x11
}
\description{
Calculates Q-matrix for bivariate profile likelihoods for Ploidevol model
}

