% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_chromevolM3.R
\name{Q_chromevolM3}
\alias{Q_chromevolM3}
\title{Calculates Q-matrix for chromevol M3 model (beta version)}
\usage{
Q_chromevolM3(log.theta, size)
}
\arguments{
\item{log.theta}{vector of size 4 indicating parameters in ln scale for chromevol M3 the order of the parameters is (lambda, mu, rho, delta)}

\item{size}{Maximum number of chromosomes in the sample (recommended no more than 40)}
}
\value{
Q a sparse matrix of size (size+10)
}
\description{
Calculates Q-matrix for chromevol M3 model (beta version)
}
\details{
Q_chomevolM3 determines the Q-matrix for a model of chromosome number change with duplications, dysploidy, and  demiploidy.
}

