% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_ploidevol.R
\name{Q_ploidevol}
\alias{Q_ploidevol}
\title{Calculates Q-matrix for Ploidevol  model}
\usage{
Q_ploidevol(log.theta)
}
\arguments{
\item{log.theta}{vector of size 6 with ln values of parameters (alpha, beta, delta, epsilon, rho, omega)}
}
\value{
Q a sparse matrix of size 11x11
}
\description{
Calculates Q-matrix for Ploidevol  model
}
\details{
This function specifies a Q-matrix of size 11x11 where states are ploidy values from 2x to 12x
}

