% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_reducedbichrom.R
\name{Q_reducedbichrom}
\alias{Q_reducedbichrom}
\title{Calculates a Q matrix for reduced BiChroM model}
\usage{
Q_reducedbichrom(log.pars, size, equal.param = c("rho0", "rho1"),
  location.par = 5)
}
\arguments{
\item{log.pars}{a vector of size 9 with  ln values of parameters (lambda0, lambda1, mu0, mu1, rho, q01, q10, e0, e1 )}

\item{size}{Maximum number of chromosomes in the sample (recommended no more than 50, states larger than that should be coded as 51)}

\item{equal.param}{by default "rho0"="rho1" but user can specify the parameters forced to be equal}

\item{location}{position in log.pars where the value of the equal.param is located, by default location=5}
}
\description{
Calculates a Q matrix for reduced BiChroM model
}
\details{
This function assumes that two parameter values are equal. This is useful for calculating likelihood ratio tests in BiChrom. By default the two parameters assumed equal are "rho0"="rho1" but user can specify which two parameters are the same
}

