% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_unibichrom.R
\name{Q_unibichrom}
\alias{Q_unibichrom}
\title{Calculates Q-matrix for univariate profile likelihoods from BiChroM model}
\usage{
Q_unibichrom(log.pars, log.theta0, size, param.name)
}
\arguments{
\item{log.pars}{vector of size 9 indicating values in ln of nuisance parameters (excluding one of interest for the profile likelihood)}

\item{log.theta0}{vector of size 1 indicating values parameters of interest in BiChroM. Options are (lambda0, lambda1, mu0, mu1, rho0, rho1, q01,q10, e01, e10)}

\item{size}{Maximum number of chromosomes in the sample (recommended no more than 50, states larger than that should be coded as 51)}

\item{param.name}{vector of size 1 indicating the names of parameters of interest. E.g. "rho0"}
}
\value{
Q a sparse matrix of size 2*(size+1)
}
\description{
Calculates Q-matrix for univariate profile likelihoods from BiChroM model
}

