% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_uniploidevol.R
\name{Q_uniploidevol}
\alias{Q_uniploidevol}
\title{Calculates Q-matrix for univariate profile likelihoods for Ploidevol model}
\usage{
Q_uniploidevol(log.pars, log.theta0, param.name)
}
\arguments{
\item{log.pars}{vector of size 5 indicating values in ln of nuisance parameters (excluding one of interest for the profile likelihood)}

\item{log.theta0}{vector of size 1 indicating values parameters of interest in BiChroM. Options are (alpha, beta,delta, epsilon, rho, omega)}

\item{param.name}{vector of size 1 indicating the names of parameters of interest. E.g. "epsilon"}
}
\value{
Q a sparse matrix of size 11x11
}
\description{
Calculates Q-matrix for univariate profile likelihoods for Ploidevol model
}

