% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationary_ploidevol.R
\name{stationary_ploidevol}
\alias{stationary_ploidevol}
\title{Calculates stationary distribution for root under Ploidevol model}
\usage{
stationary_ploidevol(log.parameters)
}
\arguments{
\item{log.parameters}{a vector of size 6 containing ln values for (alpha, beta, delta, epsilon, rho, omega)}
}
\value{
A probability vector of size 11 indicating probabilities of the root being ploidy 2x, 3x,..., 12x
}
\description{
Calculates stationary distribution for root under Ploidevol model
}
\author{
Rosana Zenil-Ferguson
}
\seealso{
\code{\link{negloglikelihood}}, \code{\link{Q_ploidevol}}
}

